+++ openapi/CHANGELOG.md
@@ -6,18 +6,20 @@
 
 
 
-## v0.107.25: API changes
+## v0.107.26: API changes
 
 ### Deprecated statistics APIs
 
 * The `GET /control/stats_info` HTTP API; use the new `GET
   /control/stats/config` API instead.
+
 * The `POST /control/stats_config` HTTP API; use the new `PUT
   /control/stats/config/update` API instead.
 
 ### New statistics APIs
 
 * The new `GET /control/stats/config` HTTP API.
+
 * The new `PUT /control/stats/config/update` HTTP API allows config updates.
 
 These `control/stats/config/update` and `control/stats/config` APIs accept and
@@ -35,12 +37,14 @@ return a JSON object with the following format:
 
 * The `GET /control/querylog_info` HTTP API; use the new `GET
   /control/querylog/config` API instead.
+
 * The `POST /control/querylog_config` HTTP API; use the new `PUT
   /control/querylog/config/update` API instead.
 
 ### New query log APIs
 
 * The new `GET /control/querylog/config` HTTP API.
+
 * The new `PUT /control/querylog/config/update` HTTP API allows config updates.
 
 These `control/querylog/config/update` and `control/querylog/config` APIs